Fixes spacing in Breadcrumbs components and moves Breadcrumbs directory back to components directory
